Add High brightness mode QS Tile
- This is a conditional tile and will only show if the proper
support is found in FWB and device tree.

NOTE:
Proof of Concept tile. This should actually be incorporated as
a conditional third state in the existing brightness tile to make
it more apparent to users that the manual and auto brightness
modes are overriden by this. Essentialy to indicate that it is
a third mode.

*DU Edits*
- Cleaned up tile some
- Use vector for icons
- Added toast when HSB is active to hopefully prevent bogus bugs

@nathanchance edits: I don't overlay the tile config so it's
built into this commit

Change-Id: I886d72b0e40bebeb3ed70bb6956ddd104842b7ed